バックエンドにチャットAPIルーターを追加し、フロントエンドに新しいページを実装。環境変数の設定方法を改善し、SPAのルーティングを整備。React Routerを導入し、アプリケーションの構成を整理。
