+++ frontend/package.json
@@ -13,6 +13,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-markdown": "^10.1.0",
+    "react-router-dom": "^7.6.2",
     "remark-gfm": "^4.0.1"
   },
   "devDependencies": {
