+++ frontend/src/pages/NewPage.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const NewPage: React.FC = () => {
+  return (
+    <div className="h-full overflow-y-auto">
+        <div className="p-8 text-white">
+          <h1 className="text-2xl font-bold mb-4">New Page</h1>
+          <p>
+            This is a new page.
+          </p>
+        </div>
+    </div>
+  );
+};
+
+export default NewPage;
