+++ frontend/vite.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
   },
   server: {
     proxy: {
-      '/chat': {
+      '/api': {
         target: 'http://127.0.0.1:8000',
         changeOrigin: true,
       },
